hadi lekhra inchaallah

+++ Minishell/execution/utils/heredoc_utils.c
@@ -3,25 +3,25 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ielouarr <ielouarr@student.42.fr>          +#+  +:+       +#+        */
+/*   By: moel-hai <moel-hai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/28 23:29:55 by ielouarr          #+#    #+#             */
-/*   Updated: 2025/07/14 11:57:43 by ielouarr         ###   ########.fr       */
+/*   Updated: 2025/07/16 03:00:48 by moel-hai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../Minishell.h"
 
-void	ft_file(t_cmd *cmd, int index, t_data *d, int *fd)
+void	ft_file(t_cmd *cmd, t_data *d, int *fd)
 {
-	char	*filename;
-	char	*num;
-
-	num = ft_itoa(index, d);
-	filename = ft_strjoin("/private/tmp/.heredoc_", num, d);
-	cmd->heredocfilename = ft_strdup(filename, d);
-	unlink(cmd->heredocfilename);
+	cmd->heredocfilename = ft_strdup("/tmp/.heredoc", d);
+	if (*fd != -1)
+		unlink(cmd->heredocfilename);
 	*fd = open(cmd->heredocfilename, O_CREAT | O_RDWR | O_TRUNC, 0644);
+	if (cmd->fd_heredoc != -1)
+		close(cmd->fd_heredoc);
+	cmd->fd_heredoc = open(cmd->heredocfilename, O_RDONLY, 0666);
+	unlink(cmd->heredocfilename);
 }
 
 void	ft_expand_heredoc_handler(t_str *current, int fd, t_data *d)
